chore: update gps pill
