+++ src/frontend/screens/MapScreen/index.tsx
@@ -13,7 +13,6 @@ import {useDraftObservation} from '../../hooks/useDraftObservation';
 import {usePersistedDraftObservation} from '../../hooks/persistedState/usePersistedDraftObservation';
 import {usePresetsQuery} from '../../hooks/server/presets';
 import ScaleBar from 'react-native-scale-bar';
-import {useLastKnownLocation} from '../../hooks/useLastSavedLocation';
 import {TrackBottomSheet} from './TrackBottomSheet';
 import {CurrentTrackMapLayer} from './CurrentTrack/CurrrentTrackMapLayer';
 
@@ -21,7 +20,7 @@ import {useMapStyleJsonUrl} from '../../hooks/server/maps';
 import {TracksMapLayer} from './MapLayers/TracksMapLayer';
 import {assert} from '../../lib/assert';
 import {RemoteDetectionAlertsMapLayer} from './MapLayers/RemoteDetectionAlertsLayer';
-import {getLocationStatus, matchPreset} from '../../lib/utils';
+import {matchPreset} from '../../lib/utils';
 import {NativeHomeTabsNavigationProps} from '../../sharedTypes/navigation';
 import {useFocusEffect} from '@react-navigation/native';
 import {GPSPill} from '../../sharedComponents/GPSPill';
@@ -55,10 +54,6 @@ export const MapScreen = ({
   const {newDraft} = useDraftObservation();
   const {navigate} = useNavigationFromHomeTabs();
   const location = useLocationState(store => store.throttledMapLocation);
-  const locationProviderStatus = useLocationState(
-    store => store.providerStatus,
-  );
-  const savedLocation = useLastKnownLocation();
   const coords = location && getCoords(location);
   const {isTracking} = useTracking();
   const {data: styleUrl} = useMapStyleJsonUrl();
@@ -114,11 +109,7 @@ export const MapScreen = ({
         }}>
         <Mapbox.Camera
           defaultSettings={{
-            centerCoordinate: coords
-              ? coords
-              : savedLocation.data
-                ? getCoords(savedLocation.data)
-                : [0, 0],
+            centerCoordinate: coords || [0, 0],
             zoomLevel: zoom,
           }}
           centerCoordinate={following ? coords : undefined}
@@ -146,15 +137,7 @@ export const MapScreen = ({
       </Mapbox.MapView>
       <View style={styles.bottomContainer}>
         <View style={{flex: 1, alignItems: 'center'}}>
-          <GPSPill
-            {...getLocationStatus({
-              providerStatus: locationProviderStatus,
-              location: location,
-            })}
-            testID="MAP.gps-pill"
-            accessibilityLabel="Open GPS Modal."
-            onPress={() => navigation.navigate('GpsModal')}
-          />
+          <GPSPill onPress={() => navigation.navigate('GpsModal')} />
         </View>
 
         <TouchableOpacity
